Movimiento de naat
- Se movió contenido de naat del modulo de cyd al modulo de naat
- Se bloqueo el menu de crear participantes de naat, ya no se utiliza
