+++ src/apps/cyd/menus.py
@@ -43,12 +43,6 @@ cyd_control_academico_children = (
         reverse_lazy("control_academico_grupo"),
         weight=12,
         icon="fa-book"),
-        
-    ViewMenuItem(
-        "Tareas Naat",
-        reverse_lazy("participantes_tareas_naat"),
-        weight=12,
-        icon="fa-tasks"),
         )
 
 
@@ -81,12 +75,6 @@ cyd_informes_children = (
         group="cyd_admin,cyd_informes",
         icon="fa-check-circle"),
     
-    ViewMenuItem(
-        "Informe avances Naat",
-        reverse_lazy("informe_naat_participantes"),
-        weight=12,
-        group="cyd_admin,cyd_informes,cyd_capacitador_informe",
-        icon="fa-share-alt"),
 
     ViewMenuItem(
         "Capacitadores",
