+++ src/apps/cyd/views.py
@@ -1902,6 +1902,7 @@ class SubirControlAcademicoExcel(views.APIView):
             status=status.HTTP_200_OK
             )
 
+
 class ImportarParticipantesNaat(View):
     """
     Vista para importar participantes desde NAAT Mobile. Recibe un código UDI de escuela como parámetro GET, 
@@ -1995,170 +1996,3 @@ class ImportarParticipantesNaat(View):
             return JsonResponse({'error': 'El servidor de Naat devolvió una respuesta inválida.'}, status=500)
         except Exception as e:
             return JsonResponse({'error': 'Error desconocido: {}'.format(str(e))}, status=500)
-
-
-class TareasNaat(LoginRequiredMixin, GroupRequiredMixin, FormView):
-    template_name = "cyd/naat/tareas_naat.html"    
-    form_class = cyd_f.ParticipanteNaatTareasForm 
-    group_required = [u"cyd", u"cyd_capacitador", u"cyd_admin", ]
-        
-
-class TareasNaatApi(views.APIView):
-    def get(self, request, *args, **kwargs):
-        sede = request.GET.get('sede', '').strip()
-        if not sede:
-            return JsonResponse({'error': 'Debes proporcionar un código UDI.'}, status=400)
-        try:
-            sede_obj = cyd_m.Sede.objects.get(id=sede)
-            escuelas = sede_obj.get_escuelas()
-            udis_escuela = [escuela.codigo for escuela in escuelas]
-            udis_listado = ",".join(udis_escuela)
-
-        except cyd_m.Sede.DoesNotExist:
-            return JsonResponse({'error': 'Sede no encontrada.'}, status=404)
-
-        url = getattr(settings, 'NAAT_URL_APIS')
-        url_base_naat = getattr(settings, 'NAAT_URL_BASE')
-        token = getattr(settings, 'NAAT_TOKEN', '')
-        endpoint = 'get_maestros_tareas'
-        url = "{}{}".format(url, endpoint)
-        headers= {
-            'API-KEY': token
-        }
-        try:
-            response = requests.get(url, params={'udi': udis_listado}, headers=headers)
-            participantes_naat = response.json()
-            maestros_agrupados = {}
-
-            for item in participantes_naat:
-                dpi = item.get('usuario_dpi')
-                
-                if dpi not in maestros_agrupados:
-                    try:
-                        maestro = cyd_m.Participante.objects.get(dpi=dpi, activo=True)
-                    except cyd_m.Participante.DoesNotExist:
-                        maestro = None
-
-                    maestros_agrupados[dpi] = {
-                        'nombre': maestro.nombre if maestro else item.get('usuario_nombre', ''),
-                        'apellido': maestro.apellido if maestro else item.get('usuario_apellido', ''),
-                        'url_participante': maestro.get_absolute_url() if maestro else "#",
-                        'registrado': True if maestro else False,
-                        'tareas': {} 
-                    }
-                nombre_tarea = item.get('nombre')
-                ruta_completa = url_base_naat + item.get('ruta', '') 
-                maestros_agrupados[dpi]['tareas'][nombre_tarea] = ruta_completa
-
-            datos_finales = list(maestros_agrupados.values())
-
-            return JsonResponse(datos_finales, safe=False, status=200)
-
-        except requests.exceptions.Timeout:
-            return JsonResponse({'error': 'El servidor tardo demasiado en responder.'}, status=504)
-        except requests.exceptions.ConnectionError:
-            return JsonResponse({'error': 'No se pudo conectar con el servidor Naat.'}, status=502)
-        except ValueError:
-            return JsonResponse({'error': 'El servidor de Naat devolvió una respuesta inválida.'}, status=500)
-        except Exception as e:
-            return JsonResponse({'error': 'Error desconocido: {}'.format(str(e))}, status=500)
-
-
-
-class InformeNaatParticipanteView(LoginRequiredMixin, GroupRequiredMixin, FormView):
-    template_name = "cyd/informeUsuariosNaat.html"    
-    form_class = cyd_f.InformeNaatParticipanteForm 
-    group_required = [u"cyd", u"cyd_capacitador", u"cyd_admin", ]
-
-
-class InformeNaatParticipanteAPI(View):
-    """
-    Vista para importar participantes desde NAAT Mobile. Recibe un código UDI de escuela como parámetro GET, 
-    consulta la API de NAAT, y devuelve un JSON con la información de los maestros asociados a esa escuela.
-    """
-    def get(self, request, *args, **kwargs):
-        sede = request.GET.get('sede').strip()
-        fecha_inicio = request.GET.get('fecha_inicio', '')
-        fecha_fin = request.GET.get('fecha_fin', '')
-        if not (sede or fecha_inicio or fecha_fin):
-            return JsonResponse({'error': 'Debes proporcionar almenos un campo para filtrar.'}, status=400)
-        try:
-            sede_obj = cyd_m.Sede.objects.get(id=sede)
-            escuelas = sede_obj.get_escuelas()
-            udis_escuela = [escuela.codigo for escuela in escuelas]
-            udis_listado = ",".join(udis_escuela)
-        except:
-            udis_listado = ""
-
-
-        url = getattr(settings, 'NAAT_URL_APIS')
-        token = getattr(settings, 'NAAT_TOKEN', '')
-        endpoint = 'get_estadisticas_api'
-        url = "{}{}".format(url, endpoint)
-
-        headers= {
-            'API-KEY': token
-        }
-        parametros = {
-            'udi': udis_listado,
-            'fecha_inicio': fecha_inicio,
-            'fecha_fin': fecha_fin
-        }
-
-        try:
-            response = requests.get(url, params=parametros, headers=headers)
-            participantes_naat = response.json()
-            listado = []
-
-            for par in participantes_naat.get('data', []):
-                dpi = par.get('dpi', {})
-                try:
-                    participante = cyd_m.Participante.objects.get(dpi=dpi, activo=True)
-                except Exception as e:
-                    participante = None
-
-                calif_lista = par.get('calificaciones', [])
-                mundos_data = { c.get('mundo_numero'): c for c in calif_lista }
-
-                data = {
-                    "nombre": participante.nombre if participante else par.get('nombre'),
-                    'apellido': participante.apellido if participante else par.get('apellido'),
-                    'url_participante': participante.get_absolute_url() if participante else None,
-                    'udi': par.get('udi', {}),
-                    'municipio': par.get('municipio_nombre', {}),
-                    'departamento': par.get('departamento_nombre', {}),
-                    'puesto': par.get('puesto_nombre', {}),
-                    'genero': par.get('genero', {}),
-
-                    "mundo1": mundos_data.get(1, {}).get('porcentaje_total', 0),
-                    "mundo2": mundos_data.get(2, {}).get('porcentaje_total', 0),
-                    "mundo3": mundos_data.get(3, {}).get('porcentaje_total', 0),
-                    "mundo4": mundos_data.get(4, {}).get('porcentaje_total', 0),
-                    "mundo5": mundos_data.get(5, {}).get('porcentaje_total', 0),
-                    "mundo6": mundos_data.get(6, {}).get('porcentaje_total', 0),
-                    "mundo7": mundos_data.get(7, {}).get('porcentaje_total', 0),
-                    "mundo8": mundos_data.get(8, {}).get('porcentaje_total', 0),
-
-                    #PC = prueba de conocimiento
-                    "mundo1_pc_nota_inicial": mundos_data.get(1, {}).get('datos_prueba', {}).get('resultado', "-"),
-                    "mundo1_pc_nota_final": mundos_data.get(1, {}).get('datos_prueba', {}).get('resultado_final', "-"),
-                    "mundo2_pc_nota_inicial": mundos_data.get(2, {}).get('datos_prueba', {}).get('resultado', "-"),
-                    "mundo2_pc_nota_final": mundos_data.get(2, {}).get('datos_prueba', {}).get('resultado_final', "-"),
-                    "mundo3_pc_nota_inicial": mundos_data.get(3, {}).get('datos_prueba', {}).get('resultado', "-"),
-                    "mundo3_pc_nota_final": mundos_data.get(3, {}).get('datos_prueba', {}).get('resultado_final', "-"),
-                    "mundo4_pc_nota_inicial": mundos_data.get(4, {}).get('datos_prueba', {}).get('resultado', "-"),
-                    "mundo4_pc_nota_final": mundos_data.get(4, {}).get('datos_prueba', {}).get('resultado_final', "-"),
-                }
-                listado.append(data)
-
-            return JsonResponse(listado, safe=False, status=200)
-
-        except requests.exceptions.Timeout:
-            return JsonResponse({'error': 'El servidor tardó demasiado en responder.'}, status=504)
-        except requests.exceptions.ConnectionError:
-            return JsonResponse({'error': 'No se pudo conectar con el servidor Naat.'}, status=502)
-        except ValueError:
-            return JsonResponse({'error': 'El servidor de Naat devolvió una respuesta inválida.'}, status=500)
-        except Exception as e:
-            return JsonResponse({'error': 'Error con la conexión', 'details': str(e)}, status=500)
-
