+++ src/apps/naat/menus.py
@@ -4,11 +4,11 @@ from apps.main.menus import ViewMenuItem
 
 # Naat
 naat_children = (
-    ViewMenuItem(
-        "Crear participante",
-        reverse_lazy("participante_naat_add"),
-        weight=30,
-        icon="fa-user"),
+    # ViewMenuItem(
+    #     "Crear participante",
+    #     reverse_lazy("participante_naat_add"),
+    #     weight=30,
+    #     icon="fa-user"),
     ViewMenuItem(
         "Calendario",
         reverse_lazy("sesion_naat_calendar"),
@@ -18,7 +18,19 @@ naat_children = (
         "Procesos de Naat",
         reverse_lazy("proceso_naat_list"),
         weight=20,
-        icon="fa-cubes"),)
+        icon="fa-cubes"),
+    ViewMenuItem(
+        "Tareas Naat",
+        reverse_lazy("participantes_tareas_naat"),
+        weight=12,
+        icon="fa-tasks"),
+    ViewMenuItem(
+        "Informe avances Naat",
+        reverse_lazy("informe_naat_participantes"),
+        weight=12,
+        group="cyd_admin,cyd_informes,cyd_capacitador_informe",
+        icon="fa-share-alt"),    
+        )
 
 Menu.add_item(
     "user",
@@ -28,4 +40,5 @@ Menu.add_item(
         weight=10,
         icon="fa-sun-o",
         group="naat",
-        children=naat_children))
+        children=naat_children)
+        )
