+++ src/templates/naat/informeUsuariosNaat.html
@@ -87,7 +87,7 @@
 {% endblock content %}
 {% block extra_js %}
 {% include "base/dataTables.html" %}
-<script src="{% static "js/extrajs/cyd.js" %}"></script>
+<script src="{% static "js/extrajs/naat.js" %}"></script>
 <script>
 	var  informeNaat = new informeParticipantesNaat();
