+++ src/templates/naat/tareas_naat.html
@@ -71,7 +71,7 @@
 {% endblock content %}
 
 {% block extra_js %}
-<script src="{% static "js/extrajs/cyd.js" %}"></script>
+<script src="{% static "js/extrajs/naat.js" %}"></script>
 <script>
 	var  nuevo = new tareasNaat();
 </script>
